Fix tests and go vet warnings
The GitHub lib's API has changed a bit, requiring the int -> int64 change. Also
fixed some new `go vet` warnings while at it.

+++ github_review_helper_suite_test.go
@@ -40,7 +40,7 @@ const (
 
 var (
 	repository = &github.Repository{
-		ID: github.Int(repositoryID),
+		ID: github.Int64(repositoryID),
 		Owner: &github.User{
 			Login: github.String(repositoryOwner),
 		},
